+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - fixed #176 - Game paused animation properly translates now
 - fixed #189 - EMP stat on new mook sheet won't modify
 - fixed #192 - Double quotes break macros for weapon names
+- fixed #195 - 'Has Suppressive Fire' option doesn't seem to do anything
 
 # Version: 0.74.1 (Hot Fix) |  Date:  2021-04-25
 **Bug Fix**