+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added "Unarmed" weapon type with optional (on by default) automatic damage determination based on BODY
 
 # DEV WIP BUG FIXES
+- fixed #176 - Game paused animation properly translates now
 - fixed #189 - EMP stat on new mook sheet won't modify
 - fixed #192 - Double quotes break macros for weapon names