+++ CHANGELOG.md
@@ -5,9 +5,11 @@
 - Added functionality to automatically resize the character and item sheets
 - added Debug Elements setting for developers
 - removed hp, humanity, and emp calculations for mooks
+- Added "Unarmed" weapon type with optional (on by default) automatic damage determination based on BODY
 
 # DEV WIP BUG FIXES
 - fixed #176 - Game paused animation properly translates now
+- fixed #187 - Item icons are now resizing correctly
 - fixed #189 - EMP stat on new mook sheet won't modify
 - fixed #192 - Double quotes break macros for weapon names